Added a couple modules
Added a hellow world module and a parameters module.
the hello world module was mostly to just play around with doing out-of-tree modules with yocto and really pretty easy so far mostly cause that was a super simple module.
The next was `paramswithtiming` which was just variable delayed print statement. I really didn't want to do this all in the init function as it didn't seem like having an init stay open forever seemed right so I did it with a workqueue although in the future I'd like to implement with a timer also.